move new tx selection algorithm into thread worker
